Added html

+++ install.ps1
@@ -1,3 +1,6 @@
-py -m venv ./virt
+if (!Test-Path./virt) {
+    py -m venv ./virt
+    virt/Scripts/activate
+    py -m pip install -r .\requirements.txt
+}
 virt/Scripts/activate
-py -m pip install -r .\requirements.txt
